requirements.txt
streamlit
numpy
scikit-learn
Pillow>=10.0


# local  package
-e .